# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marktr API — news, stocks, and investment simulation
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Marktr's editorial category
 */
export type Category = (typeof Category)[keyof typeof Category];

export const Category = {
  celebrity: "celebrity",
  tech: "tech",
  government: "government",
  sports: "sports",
} as const;

export interface Story {
  articleId: string;
  title: string;
  description?: string;
  link: string;
  mediaUrl?: string;
  publishedDate: string;
  source: string;
  category: Category;
  /** Primary stock ticker mentioned or implied (null if none) */
  ticker?: string;
  companyName?: string;
  /** One-line financial insight written by Claude */
  insight: string;
  /** Plain-language financial explanation written by Claude */
  explanation: string;
}

export interface StoriesResponse {
  date: string;
  totalCount: number;
  stories: Story[];
}

export interface StockPricePoint {
  date: string;
  close: number;
}

export interface StockHistory {
  ticker: string;
  currency?: string;
  range: string;
  points: StockPricePoint[];
  latestPrice?: number;
  previousClose?: number;
  cachedAt: string;
}

export interface SimulateRequest {
  ticker: string;
  amount: number;
  /** ISO date YYYY-MM-DD when the investment was made */
  startDate: string;
}

export interface SimulateResponse {
  ticker: string;
  startDate: string;
  amount: number;
  startPrice: number;
  currentPrice: number;
  sharesPurchased: number;
  currentValue: number;
  profitLoss: number;
  percentChange: number;
}

export type IngestionStatusState =
  (typeof IngestionStatusState)[keyof typeof IngestionStatusState];

export const IngestionStatusState = {
  idle: "idle",
  running: "running",
  success: "success",
  error: "error",
} as const;

export interface IngestionStatus {
  date?: string;
  state: IngestionStatusState;
  storyCount?: number;
  startedAt?: string;
  finishedAt?: string;
  message?: string;
}

export interface ErrorResponse {
  message: string;
  code?: string;
}

export type ListStoriesParams = {
  category?: Category;
  /**
   * @minimum 1
   * @maximum 200
   */
  limit?: number;
};

export type GetStockHistoryParams = {
  range?: GetStockHistoryRange;
};

export type GetStockHistoryRange =
  (typeof GetStockHistoryRange)[keyof typeof GetStockHistoryRange];

export const GetStockHistoryRange = {
  "1mo": "1mo",
  "3mo": "3mo",
  "6mo": "6mo",
  "1y": "1y",
  "5y": "5y",
} as const;