# lib/api-zod/src/generated/types/ingestionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marktr API — news, stocks, and investment simulation
 * OpenAPI spec version: 0.1.0
 */
import type { IngestionStatusState } from "./ingestionStatusState";

export interface IngestionStatus {
  date?: string;
  state: IngestionStatusState;
  storyCount?: number;
  startedAt?: string;
  finishedAt?: string;
  message?: string;
}