# lib/api-zod/src/generated/types/simulateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marktr API — news, stocks, and investment simulation
 * OpenAPI spec version: 0.1.0
 */

export interface SimulateRequest {
  ticker: string;
  amount: number;
  /** ISO date YYYY-MM-DD when the investment was made */
  startDate: string;
}