# lib/api-zod/src/generated/types/simulateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marktr API — news, stocks, and investment simulation
 * OpenAPI spec version: 0.1.0
 */

export interface SimulateResponse {
  ticker: string;
  startDate: string;
  amount: number;
  startPrice: number;
  currentPrice: number;
  sharesPurchased: number;
  currentValue: number;
  profitLoss: number;
  percentChange: number;
}