# lib/api-zod/src/generated/types/stockHistory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marktr API — news, stocks, and investment simulation
 * OpenAPI spec version: 0.1.0
 */
import type { StockPricePoint } from "./stockPricePoint";

export interface StockHistory {
  ticker: string;
  currency?: string;
  range: string;
  points: StockPricePoint[];
  latestPrice?: number;
  previousClose?: number;
  cachedAt: string;
}