# lib/api-zod/src/generated/types/storiesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marktr API — news, stocks, and investment simulation
 * OpenAPI spec version: 0.1.0
 */
import type { Story } from "./story";

export interface StoriesResponse {
  date: string;
  totalCount: number;
  stories: Story[];
}