# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marktr API — news, stocks, and investment simulation
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface Story {
  articleId: string;
  title: string;
  description?: string;
  link: string;
  mediaUrl?: string;
  publishedDate: string;
  source: string;
  category: Category;
  /** Primary stock ticker mentioned or implied (null if none) */
  ticker?: string;
  companyName?: string;
  /** One-line financial insight written by Claude */
  insight: string;
  /** Plain-language financial explanation written by Claude */
  explanation: string;
}